# management-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Сервис аутентификации",
        "title": "Auth-Service",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "2.0"
    },
    "host": "localhost:8082",
    "paths": {
        "/programs": {
            "get": {
                "description": "Получить все программы тренировок пользователя",
                "tags": [
                    "Программы тренировок"
                ],
                "summary": "Получить все программы тренировок пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Логин пользователя",
                        "name": "X-User-Login",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Программы тренировок",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal.ProgramDb"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            },
            "post": {
                "description": "Создать программу тренировок",
                "tags": [
                    "Программы тренировок"
                ],
                "summary": "Создать программу тренировок",
                "parameters": [
                    {
                        "description": "Данные для создания программы",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.CreateProgramReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Программа тренировок создана"
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            }
        },
        "/programs/trainings": {
            "post": {
                "description": "Добавить тренировку в программу тренировок",
                "tags": [
                    "Программы тренировок"
                ],
                "summary": "Добавить тренировку в программу тренировок",
                "parameters": [
                    {
                        "description": "Данные для добавления тренировки в программу",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.AddTrainingToProgramReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировка добавлена в программу"
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            },
            "delete": {
                "description": "Удалить тренировку из программы тренировок",
                "tags": [
                    "Программы тренировок"
                ],
                "summary": "Удалить тренировку из программы тренировок",
                "parameters": [
                    {
                        "description": "Данные для удаления тренировки из программы",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.DeleteTrainingToProgramReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировка удалена из программы"
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            }
        },
        "/programs/{id}": {
            "get": {
                "description": "Получить программу тренировок",
                "tags": [
                    "Программы тренировок"
                ],
                "summary": "Получить программу тренировок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Логин пользователя",
                        "name": "X-User-Login",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id программы тренировок",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Программа тренировок",
                        "schema": {
                            "$ref": "#/definitions/internal.ProgramDb"
                        }
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            },
            "delete": {
                "description": "Удалить программу тренировок",
                "tags": [
                    "Программы тренировок"
                ],
                "summary": "Удалить программу тренировок",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Логин пользователя",
                        "name": "X-User-Login",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "id программы тренировок",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно удалили"
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            }
        },
        "/programs/{id}/trainings": {
            "get": {
                "description": "Получить тренировки конкретной программы",
                "tags": [
                    "Программы тренировок"
                ],
                "summary": "Получить тренировки конкретной программы",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id программы тренировок",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Логин пользователя",
                        "name": "X-User-Login",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировки программы",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal.TrainingDb"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            }
        },
        "/trainings": {
            "get": {
                "description": "Получить все тренировки пользователя",
                "tags": [
                    "Тренировки"
                ],
                "summary": "Получить все тренировки пользователя",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Логин пользователя",
                        "name": "login",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировки пользователя",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal.TrainingDb"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            },
            "post": {
                "description": "Создать тренировку",
                "tags": [
                    "Тренировки"
                ],
                "summary": "Создать тренировку",
                "parameters": [
                    {
                        "description": "Данные для создания тренировки",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.TrainingDb"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировка создана"
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            }
        },
        "/trainings/{id}": {
            "get": {
                "description": "Получить тренировку по id",
                "tags": [
                    "Тренировки"
                ],
                "summary": "Получить тренировку по id",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id тренировки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Логин пользователя",
                        "name": "login",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировка",
                        "schema": {
                            "$ref": "#/definitions/internal.TrainingDb"
                        }
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            },
            "delete": {
                "description": "Удалить тренировку",
                "tags": [
                    "Тренировки"
                ],
                "summary": "Удалить тренировку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id тренировки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Логин пользователя",
                        "name": "X-User-Login",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировка удалена"
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            },
            "patch": {
                "description": "Обновить тренировку",
                "tags": [
                    "Тренировки"
                ],
                "summary": "Обновить тренировку",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id тренировки",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для обновления тренировки",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.TrainingDb"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Тренировка обновлена"
                    },
                    "400": {
                        "description": "Ошибка"
                    }
                }
            }
        }
    },
    "definitions": {
        "internal.AddTrainingToProgramReq": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer"
                },
                "program_id": {
                    "type": "integer"
                },
                "training": {
                    "type": "integer"
                }
            }
        },
        "internal.CreateProgramReq": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "user_login": {
                    "type": "string"
                }
            }
        },
        "internal.DeleteTrainingToProgramReq": {
            "type": "object",
            "properties": {
                "day": {
                    "type": "integer"
                },
                "program_id": {
                    "type": "integer"
                },
                "training": {
                    "type": "integer"
                }
            }
        },
        "internal.Exercise": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "sets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal.Set"
                    }
                }
            }
        },
        "internal.ProgramDb": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "userLogin": {
                    "type": "string"
                }
            }
        },
        "internal.Set": {
            "type": "object",
            "properties": {
                "reps": {
                    "type": "integer"
                },
                "weight": {
                    "type": "number"
                }
            }
        },
        "internal.TrainingDb": {
            "type": "object",
            "properties": {
                "exercises": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal.Exercise"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "userLogin": {
                    "type": "string"
                }
            }
        }
    }
}